Conditionally display the NavBar based on different screen sizes
