+++ src/components/layout/NavBar/NavBar.jsx
@@ -1,53 +1,18 @@
-import homeIcon from "../../../assets/icons/home.svg";
-import moreIcon from "../../../assets/icons/dotsVerticalMore.svg";
-import { useNavigate, useLocation } from "react-router-dom";
-import { useUserRole } from "../../UserRoleContext/UserRoleContext";
+import NavBarMobile from "./NavBarMobile";
+import NavBarDesktop from "./NavBarDesktop";
 import "./NavBar.css";
 
-const NavBar = () => {
-  const navigate = useNavigate();
-  const location = useLocation();
-  const { userRole } = useUserRole();
-
-  const isHomepageActive = location.pathname === "/patient/homepage" || location.pathname === "/hp/homepage";
-
+function NavBar() {
   return (
-    <nav className="navbar">
-      {userRole === "patient" ? (
-        <>
-          <button
-            className={`btn homeNav ${isHomepageActive ? "active" : ""}`}
-            onClick={() => navigate("/patient/homepage")}
-          >
-            <img src={homeIcon} alt="Home Icon" className="homeNavIcon" />
-            Home
-          </button>
-
-          <button className="btn moreNav">
-            {/* onclick open menu */}
-            <img src={moreIcon} alt="More Icon" className="moreNavIcon" />
-            More
-          </button>
-        </>
-      ) : (
-        <>
-          <button
-           className={`btn homeNav ${isHomepageActive ? "active" : ""}`}
-            onClick={() => navigate("/hp/homepage")}
-          >
-            <img src={homeIcon} alt="Home Icon" className="homeNavIcon" />
-            Home
-          </button>
-
-          <button className="btn moreNav">
-            {/* onclick open menu */}
-            <img src={moreIcon} alt="More Icon" className="moreNavIcon" />
-            More
-          </button>
-        </>
-      )}
-    </nav>
+    <>
+      <nav className="NavBar NavBarMobile">
+        <NavBarMobile />
+      </nav>
+      <nav className="NavBar NavBarDesktop">
+        <NavBarDesktop />
+      </nav>
+    </>
   );
-};
+}
 
 export default NavBar;
