+++ src/components/layout/NavBar/NavBarDesktop.css
@@ -3,7 +3,8 @@
     width: 100%;
     justify-content: space-between;
     align-items: center;
-    margin: 0 50px;
+ /*    margin: 0 50px; */
+    background-color: var(--primary-color);
 }
 
 .logoDesktop {
