+++ src/components/layout/NavBar/NavBarMobile.css
@@ -1,0 +1,78 @@
+.navbar {
+    display: flex;
+    position: fixed;
+    bottom: 0;
+    left: 0;
+    right: 0;
+    width: 100%;
+    height: 90px;
+    padding-left: 70px;
+    padding-right: 70px;
+    justify-content: space-between;
+    align-items: center;
+    background: #e3eac4;
+    box-sizing: border-box;
+  }
+  
+  .btn {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    background: #e3eac4;
+    width: 70px;
+    height: 70px;
+    border: none;
+    box-shadow: none;
+    font-size: 0.75rem;
+    cursor: pointer;
+  }
+  
+  .btn:hover {
+    background-color:  var(--hover-active-color);
+    border-radius: 0.63rem;
+    padding: 5px 10px;
+    transition: background-color 0.3s ease;
+  }
+  
+  .homeNavIcon {
+    width: 40px;
+    height: 40px;
+    flex-shrink: 0;
+    aspect-ratio: 1/1;
+  }
+  
+  .moreNavIcon {
+    width: 40px;
+    height: 40px;
+    flex-shrink: 0;
+    aspect-ratio: 1/1;
+  }
+  
+  .navbar .homeNav.active {
+    background-color: var(--hover-active-color);
+    border-radius: 0.63rem;
+  }
+  
+  @media (min-width: 600px) and (max-width: 1024px) {
+    .homeNavIcon {
+      width: 60px;
+      height: 60px;
+    }
+    .moreNavIcon {
+      width: 60px;
+      height: 60px;
+    }
+  
+    .btn {
+      font-size: 1.5rem;
+      width: 120px;
+      height: 120px;
+    }
+    .navbar {
+      height: 140px;
+      padding-left: 150px;
+      padding-right: 150px;
+    }
+  }
+
