+++ src/routes/patient/PatientHomepage/PatientHomepage.css
@@ -42,5 +42,4 @@
     font-size: 1rem;
     justify-self:left;
     font-family: "Inter", sans-serif;
-  
 }
