+++ src/routes/patient/PatientHomepage/PatientHomepage.jsx
@@ -1,6 +1,7 @@
 import "./PatientHomepage.css";
 import homeIcon from "../../../assets/icons/home.svg";
 import ButtonsPatientHomepage from "../../../components/ui/ButtonsPatientHomepage/ButtonsPatientHomepage";
+import NavBarMobileModal from "../../../components/layout/NavBar/NavBarMobileModal/NavBarMobileModal";
 
 function PatientHomepage() {
   return (
@@ -13,6 +14,7 @@ function PatientHomepage() {
         </div>
       </header>
       <ButtonsPatientHomepage />
+      <NavBarMobileModal />
     </div>
   );
 }
